requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-import-export==4.1.1
django-leaflet==0.30.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
tablib==3.5.0
